layout/style/test/test_bug580685.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=580685
-->
<head>
  <title>Test for Bug 580685</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=580685">Mozilla Bug 580685</a>
<p id="display">
  <iframe id="f" srcdoc="<body style='outline-offset: 1rem'>">
  </iframe>
</p>
<div id="content" style="display: none">
  
</div>
<pre id="test">
<script type="application/javascript">

/** Test for Bug 580685 **/
SimpleTest.waitForExplicitFinish()
addLoadEvent(function() {
  var doc = $("f").contentDocument;
  var b = doc.body;
  doc.removeChild(doc.documentElement);
  is(doc.defaultView.getComputedStyle(b).outlineOffset,
     doc.defaultView.getComputedStyle(b).fontSize,
     "1rem did not compute correctly");
  SimpleTest.finish();
});




</script>
</pre>
</body>
</html>
